Break up tests into their categories instead of putting them all in one L1 file
