+++ src/Test/L1/Worker/VariableL1Tests.cs
@@ -1,0 +1,77 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using Microsoft.TeamFoundation.DistributedTask.Pipelines;
+using Microsoft.TeamFoundation.DistributedTask.WebApi;
+using System.Linq;
+using System.Threading.Tasks;
+using Xunit;
+
+namespace Microsoft.VisualStudio.Services.Agent.Tests.L1.Worker
+{
+    public class VariableL1Tests : L1TestBase
+    {
+        [Fact]
+        [Trait("Level", "L1")]
+        [Trait("Category", "Worker")]
+        public async Task SetVariable_ReadVariable()
+        {
+            // Arrange
+            var message = LoadTemplateMessage();
+            // Remove all tasks
+            message.Steps.Clear();
+            // Add variable setting tasks
+            message.Steps.Add(CreateScriptTask("echo \"##vso[task.setvariable variable=testVar]b\""));
+            message.Steps.Add(CreateScriptTask("echo TestVar=$(testVar)"));
+            message.Variables.Add("testVar", "a");
+
+            // Act
+            var results = await RunWorker(message);
+
+            // Assert
+            AssertJobCompleted();
+            Assert.Equal(TaskResult.Succeeded, results.Result);
+
+            var steps = GetSteps();
+            Assert.Equal(4, steps.Count()); // Init, CmdLine, CmdLine, Finalize
+            var outputStep = steps[2];
+            var log = GetTimelineLogLines(outputStep);
+
+            Assert.True(log.Where(x => x.Contains("TestVar=b")).Count() > 0);
+        }
+
+        // Enable this test when read only variable enforcement is added
+        /*[Fact]
+        [Trait("Level", "L1")]
+        [Trait("Category", "Worker")]
+        public async Task Readonly_Variables()
+        {
+            // Arrange
+            var message = LoadTemplateMessage();
+            // Remove all tasks
+            message.Steps.Clear();
+            // Add a normal step and one that only runs on failure
+            message.Steps.Add(CreateScriptTask("echo ##vso[task.setvariable variable=system]someothervalue"));
+            var alwayStep = CreateScriptTask("echo SystemVariableValue=$(system)");
+            alwayStep.Condition = "always()";
+            message.Steps.Add(alwayStep);
+
+            // Act
+            var results = await RunWorker(message);
+
+            // Assert
+            AssertJobCompleted();
+            Assert.Equal(TaskResult.Succeeded, results.Result);
+
+            var steps = GetSteps();
+            Assert.Equal(4, steps.Count()); // Init, CmdLine, CmdLine, Finalize
+
+            var failToSetStep = steps[1];
+            Assert.Equal(TaskResult.Failed, failToSetStep.Result);
+
+            var outputStep = steps[2];
+            var log = GetTimelineLogLines(outputStep);
+            Assert.True(log.Where(x => x.Contains("SystemVariableValue=build")).Count() == 1);
+        }*/
+    }
+}
